Renamed ReportMetric to MetricReporter
