+++ special-pages/pages/duckplayer/src/index.js
@@ -4,7 +4,7 @@ import { createSpecialPageMessaging } from '../../../shared/create-special-page-
 import { init } from '../app/index.js';
 import { initStorage } from './storage.js';
 import '../../../shared/live-reload.js';
-import { ReportMetric, EXCEPTION_KIND_MESSAGING_ERROR } from '../../../shared/report-metric.js';
+import { MetricsReporter, EXCEPTION_KIND_MESSAGING_ERROR } from '../../../shared/metrics-reporter.js';
 
 export class DuckplayerPage {
     /**
@@ -13,7 +13,7 @@ export class DuckplayerPage {
     constructor(messaging, injectName) {
         this.messaging = createTypedMessages(this, messaging);
         this.injectName = injectName;
-        this.metrics = new ReportMetric(messaging);
+        this.metrics = new MetricsReporter(messaging);
     }
 
     /**
