+++ special-pages/unit-test/metrics-reporter.mjs
@@ -1,6 +1,6 @@
 import { describe, it, mock, beforeEach } from 'node:test';
 import assert from 'node:assert';
-import { ReportMetric } from '../shared/report-metric.js';
+import { MetricsReporter } from '../shared/metrics-reporter.js';
 
 describe('reportMetric', () => {
     let messaging;
@@ -16,18 +16,18 @@ describe('reportMetric', () => {
 
     it('should throw an error if messaging is not provided', () => {
         // @ts-expect-error - this is a forced error
-        assert.throws(() => new ReportMetric(null));
+        assert.throws(() => new MetricsReporter(null));
     });
 
     it('should throw an error if metricName is not provided', () => {
-        const metrics = new ReportMetric(messaging);
+        const metrics = new MetricsReporter(messaging);
         const metricParams = /** @type {any} */ ({ metricName: '', params: {} });
         assert.throws(() => metrics.reportMetric(metricParams));
         assert.strictEqual(messaging.notify.mock.callCount(), 0);
     });
 
     it('should call messaging.notify with the correct parameters', () => {
-        const metrics = new ReportMetric(messaging);
+        const metrics = new MetricsReporter(messaging);
         const metricParams = /** @type {any} */ ({ metricName: 'exception', params: { message: 'This is a test' } });
         assert.strictEqual(messaging.notify.mock.callCount(), 0);
 
@@ -44,7 +44,7 @@ describe('reportMetric', () => {
     });
 
     it('should call messaging.notify when reportException is called', () => {
-        const metrics = new ReportMetric(messaging);
+        const metrics = new MetricsReporter(messaging);
         const eventParams = /** @type {any} */ ({ message: 'This is a test', kind: 'TestError' });
         assert.strictEqual(messaging.notify.mock.callCount(), 0);
 
@@ -61,7 +61,7 @@ describe('reportMetric', () => {
     });
 
     it('should send default values when reportException is called with empty params', () => {
-        const metrics = new ReportMetric(messaging);
+        const metrics = new MetricsReporter(messaging);
         const eventParams = /** @type {any} */ ({});
         assert.strictEqual(messaging.notify.mock.callCount(), 0);
 
@@ -78,7 +78,7 @@ describe('reportMetric', () => {
     });
 
     it('should not report anything when reportExceptionWithError is called with a non-Error object', () => {
-        const metrics = new ReportMetric(messaging);
+        const metrics = new MetricsReporter(messaging);
         const eventParams = /** @type {any} */ ({ message: 'This is a test', kind: 'TestError' });
         assert.strictEqual(messaging.notify.mock.callCount(), 0);
 
@@ -87,7 +87,7 @@ describe('reportMetric', () => {
     });
 
     it('should send the error message and kind when reportExceptionWithError is called with an Error object', () => {
-        const metrics = new ReportMetric(messaging);
+        const metrics = new MetricsReporter(messaging);
         const error = new Error('This is a test');
         error.name = 'TestError';
         assert.strictEqual(messaging.notify.mock.callCount(), 0);
@@ -105,7 +105,7 @@ describe('reportMetric', () => {
     });
 
     it('should send default values when reportExceptionWithError is called with an empty error object', () => {
-        const metrics = new ReportMetric(messaging);
+        const metrics = new MetricsReporter(messaging);
         const error = new Error();
         assert.strictEqual(messaging.notify.mock.callCount(), 0);
