+++ typedoc.js
@@ -36,7 +36,7 @@ const config = {
         'special-pages/pages/new-tab/app/favorites/constants.js',
         'special-pages/pages/**/types/*.ts',
         'special-pages/shared/types/*.ts',
-        'special-pages/shared/report-metric.js',
+        'special-pages/shared/metrics-reporter.js',
     ],
     categoryOrder: ['Special Pages', 'Content Scope Scripts Integrations', 'Other'],
     out: 'docs',
